+++ src/literalizer/languages/python.py
@@ -1086,6 +1086,6 @@ class Python(metaclass=LanguageCls):
         )
 
     @cached_property
-    def call_style_config(self) -> CallStyle | None:
+    def call_style_config(self) -> CallStyle:
         """Configuration for the chosen call style."""
         return cast("CallStyle", self.call_style.value)